requirements.txt
functions-framework==3.8.2
google-cloud-bigquery==3.27.0
google-cloud-secret-manager==2.22.0
requests==2.32.3
pytest==8.3.5
pytest-cov==6.0.0
